<Sigh> Another bug with opacity slider fixed. Wouldn't respond to all methods of input (clicking, wheel), now it does. Yay.
svn:r164

+++ src/config/vidaliasettings.h
@@ -91,8 +91,8 @@ public:
   void setMaxMsgCount(int max);
 
   /* Get and set Msg Log window opacity */
-  qreal getMsgLogOpacity();
-  void setMsgLogOpacity(qreal value);
+  int getMsgLogOpacity();
+  void setMsgLogOpacity(int value);
 };
 
 #endif
